filtr http requests using  middleware and create IsAdmin middleware

+++ All Tasks of Laravel/routes/web.php
@@ -8,12 +8,13 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
-Route::get('/posts/create',[PostController::class, 'create'])->name('posts.create');
-Route::get('/posts/{post}', [PostController::class, 'show'])->name('posts.show');
-Route::get('/posts/{post}/edit',[Postcontroller::class,'edit'])->name('posts.edit');      //edit
-Route::delete('/posts/{post}',[Postcontroller::class,'destroy'])->name('posts.destroy');  //delete
-Route::post('/posts',[PostController::class, 'store'])->name('posts.store');
+Route::get('/posts', [PostController::class, 'index'])->name('posts.index')-> middleware('auth');
+Route::get('/posts/create',[PostController::class, 'create'])->name('posts.create') -> middleware(['auth', 'isAdminMiddleware']);
+Route::get('/posts/{post}', [PostController::class, 'show'])->name('posts.show')-> middleware('auth');
+Route::put('/posts/{post}',[PostController::class,'update'])->name('posts.update')-> middleware('auth');
+Route::delete('/posts/{post}',[Postcontroller::class,'destroy'])->name('posts.destroy')-> middleware('auth');  //delete
+Route::get('/posts/{post}/edit',[Postcontroller::class,'edit'])->name('posts.edit')-> middleware('auth');      //edit
+Route::post('/posts',[PostController::class, 'store'])->name('posts.store')-> middleware('auth');
 
 Auth::routes();
